Thong tin ca nhan

+++ src/main/java/com/company/quanlyshipper/controller/Khachhang.java
@@ -8,11 +8,13 @@ package com.company.quanlyshipper.controller;
 
 import com.company.quanlyshipper.model.Areas;
 import com.company.quanlyshipper.model.Customer;
+import com.company.quanlyshipper.model.Orders;
 import com.company.quanlyshipper.model.Users;
 import com.company.quanlyshipper.repo.AreasRepo;
 import com.company.quanlyshipper.service.AreaService;
 import com.company.quanlyshipper.service.CustomerService;
 import com.company.quanlyshipper.service.LoginService;
+import com.company.quanlyshipper.service.OrderService;
 import com.company.quanlyshipper.service.UserService;
 import java.net.URL;
 import java.util.List;
@@ -41,6 +43,8 @@ public class Khachhang implements Initializable {
 
     @Autowired
     private CustomerService service;
+    @Autowired
+    private OrderService orderService;
     
     @Autowired
     private AreaService areaService;
@@ -172,13 +176,26 @@ public class Khachhang implements Initializable {
                 .message("Vui lòng chọn 1 thông tin trong bảng để thực hiện thao tác")
                 .build().show();
         } else {
-           Thongbao.ThongbaoBuilder.builder()
+           boolean isHaveOrder = false;
+           for(Orders x : orderService.getAllOrder("", "")){
+               if(x.getCustomer().getId() == this.cus.getId()){
+                   isHaveOrder = true;
+               }
+           }
+           if(isHaveOrder == true){
+               Thongbao.ThongbaoBuilder.builder()
+                .message("Bạn không thể xóa khách hàng đang có đơn hàng")
+                .build().show(); 
+           }
+           if(isHaveOrder == false){
+               Thongbao.ThongbaoBuilder.builder()
                 .title("Thông tin của khách này sẽ bị xóa ra khỏi hệ thống")
                 .message("Việc làm này sẽ không thể hoàn tác , Bạn có chắc chắn muốn xóa thông tin này không ?")
                 .okAction(()-> {
                     service.deleteCus(this.cus);
                     search();
                 }).build().show(); 
+           }
         }
         
     }
